finalizando a validação de permissões

+++ RM.ApiDotNet6.Api/Controllers/PersonController.cs
@@ -2,25 +2,33 @@
 using Microsoft.AspNetCore.Mvc;
 using RM.ApiDotNet6.Application.DTOs;
 using RM.ApiDotNet6.Application.Services.Interfaces;
+using RM.ApiDotNet6.Domain.Authentication;
 using RM.ApiDotNet6.Domain.FiltersDb;
 
 namespace RM.ApiDotNet6.Api.Controllers
 {
-    [Authorize]
     [Route("api/[controller]")]
-    [ApiController]
-    public class PersonController : ControllerBase
+    public class PersonController : BaseController
     {
         private readonly IPersonService _personService;
+        private readonly ICurrentUser _currentUser;
+        private List<string> _permissionNeeded = new List<string>() { "Admin" };
+        private readonly List<string> _permissionUser;
 
-        public PersonController(IPersonService personService)
+        public PersonController(IPersonService personService, ICurrentUser currentUser)
         {
             _personService = personService;
+            _currentUser = currentUser;
+            _permissionUser = _currentUser.Permissions.Split(",").ToList() ?? new List<string>();
         }
 
         [HttpPost]
-        public async Task<ActionResult> PostAsync([FromBody] PersonDTO personDTO)
+        public async Task<IActionResult> PostAsync([FromBody] PersonDTO personDTO)
         {
+            _permissionNeeded.Add("CadastraPessoa");
+
+            if (!ValidPermission(_permissionUser, _permissionNeeded)) return Forbidden();
+
             var result = await _personService.CreateAsync(personDTO);
 
             if (!result.IsSuccess)
@@ -30,8 +38,12 @@ namespace RM.ApiDotNet6.Api.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult> GetAsync()
+        public async Task<IActionResult> GetAsync()
         {
+            _permissionNeeded.Add("BuscaPessoa");
+
+            if (!ValidPermission(_permissionUser, _permissionNeeded)) return Forbidden();
+
             var result = await _personService.GetAsync();
 
             if (!result.IsSuccess)
@@ -42,8 +54,12 @@ namespace RM.ApiDotNet6.Api.Controllers
 
         [HttpGet]
         [Route("{id}")]
-        public async Task<ActionResult> GetByIdAsync(int id)
+        public async Task<IActionResult> GetByIdAsync(int id)
         {
+            _permissionNeeded.Add("BuscaPessoa");
+
+            if (!ValidPermission(_permissionUser, _permissionNeeded)) return Forbidden();
+
             var result = await _personService.GetByIdAsync(id);
 
             if (!result.IsSuccess)
@@ -53,8 +69,12 @@ namespace RM.ApiDotNet6.Api.Controllers
         }
 
         [HttpPut]
-        public async Task<ActionResult> UpdateAsync([FromBody] PersonDTO personDTO)
+        public async Task<IActionResult> UpdateAsync([FromBody] PersonDTO personDTO)
         {
+            _permissionNeeded.Add("EditaPessoa");
+
+            if (!ValidPermission(_permissionUser, _permissionNeeded)) return Forbidden();
+
             var result = await _personService.UpdateAsync(personDTO);
 
             if (!result.IsSuccess)
@@ -65,8 +85,12 @@ namespace RM.ApiDotNet6.Api.Controllers
 
         [HttpDelete]
         [Route("{id}")]
-        public async Task<ActionResult> DeleteAsync(int id)
+        public async Task<IActionResult> DeleteAsync(int id)
         {
+            _permissionNeeded.Add("DeletaPessoa");
+
+            if (!ValidPermission(_permissionUser, _permissionNeeded)) return Forbidden();
+
             var result = await _personService.DeleteAsync(id);
 
             if (!result.IsSuccess)
@@ -77,8 +101,12 @@ namespace RM.ApiDotNet6.Api.Controllers
 
         [HttpGet]
         [Route("paged")]
-        public async Task<ActionResult> GetPagedAsync([FromQuery] PersonFilterDb request)
+        public async Task<IActionResult> GetPagedAsync([FromQuery] PersonFilterDb request)
         {
+            _permissionNeeded.Add("BuscaPessoa");
+
+            if (!ValidPermission(_permissionUser, _permissionNeeded)) return Forbidden();
+
             var result = await _personService.GetPagedAsync(request);
 
             if (!result.IsSuccess)
